配置微服务打包命令，修复eslint对于import {React} from 'react'，但不使用的问题
