+++ command.ts
@@ -1,0 +1,18 @@
+import { exec } from "child_process";
+// 项目列表
+export const siteList = ["main", "react", "vue3"];
+
+export function runCommand(command) {
+    return new Promise((resolve, reject) => {
+        const childProcess = exec(command, (error, stdout, stderr) => {
+            if (error) {
+                reject(error);
+            } else {
+                resolve({ stdout, stderr });
+            }
+        });
+
+        childProcess.stdout?.pipe(process.stdout);
+        childProcess.stderr?.pipe(process.stderr);
+    });
+}
