+++ site/react/src/App.tsx
@@ -3,7 +3,7 @@ import Nav from "./components/Nav";
 import { BrowserRouter } from "react-router-dom";
 import BaseRouter from "./router";
 
-function App() {
+function App():React.JSX.Element {
   return (
     <>
       <BrowserRouter>
