+++ site/react/src/pages/apiUse/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
 
-export default function ApiUse() {
+export default function ApiUse():React.JSX.Element {
   return <div>接口请求</div>;
 }
