+++ site/react/src/pages/lifeCycle/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
 
-export default function LifeCycle() {
+export default function LifeCycle():React.JSX.Element {
   return <div>生命周期</div>;
 }
