+++ site/react/src/pages/notFound/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Empty } from "@arco-design/web-react";
 
-export default function NotFound() {
+export default function NotFound():React.JSX.Element {
   return <Empty />;
 }
